requirements.txt
# Core dependencies
boto3==1.28.38
python-dotenv==1.0.0
requests==2.31.0

# YouTube API
google-api-python-client==2.97.0
youtube-transcript-api==0.6.1

# OpenAI
openai==1.3.5

# Twitter/X API
tweepy==4.14.0

# AWS Lambda
aws-lambda-powertools==2.23.0

# Testing
pytest==7.4.0
moto==4.2.2
pytest-mock==3.11.1

# Development
black==23.7.0
isort==5.12.0
flake8==6.1.0 